Use merge to override app of apps params
This is a test to see if there is an impact on the interpolation when
doing it on Terraform instead of letting helm do it. The purpose of this
is to see if it would be possible to use the argocd provider, that
supports only one values, instead of helm.