fix(sks): module error when keycloak is disabled
The keycloak_admin_password output can't be set when keycloak is disabled
because the kubernetes secret doesn't exist.

This solution doesn't rely on whether the secret exists or not, but on whether
keycloak is disabled in the app_of_apps_overrides variable. The only drawback
I see is that this condition will have to be changed if the default value for
apps.keycloak.enabled is ever changed (in modules/values.tmpl.yaml).