chore: remove old deployment examples
I think these examples were used to show a deployment of the DevOps Stack in different platforms. Most of the code seems to be deprecated and I find theses examples redundant with what we have been putting in the tests folder. In my view, the tests folder should become the new examples folder.