ci: hardcode App ID on central workflow
Although I would have preferred to keep this a secret, since the application is private to our organization, I decided to hardcode the App ID on the central workflow to avoid having to manage another secret on all repos besides the private key of the app. This way, in the eventuality that we need to redo the app, we won't need to change both secrets on all the repositories, so we cut the maintenance in half.